Lower test coverage
When we have this little code and working with REST API's we could lower
the coverage a bit. It is hard to unit test these kind of stuff and
mocking just to get up the coverage feels useful. It could be useful in
integration tests, but this should not be run all the time and would
make the coverage fail if we are only running the unit tests.