Fix move linked issue workflow bug
We should fetch the list with the last 100 items instead of the first
50. I think it is ordered on when they got added to the project. The
last entries will always be more important than the first ones. Also
increase the list length to 100 to allow the cards to be stale for
longer and still work with this workflow when they are being worked on.