Add compose file and update server to use secret
We have a small README on how to use these containerized files. The
compose file is using a secret with the `OPENAI_API_KEY` which will get
mounted to the docker container. In the future we want to migrate to
something like `docker-swarm`, `vault` or a cloud providers secret
manager. But for now, this works and is better than providing a
environment variable directly to the container. See
https://docs.docker.com/compose/use-secrets/ for reference about docker
secrets.

So right now for local testing we need:

- a file `openai-key.txt` with the openai API key

- run `docker compose up --build`

- test if the REST API is accessible at localhost:8000/docs

Note that these files are generated with the help of `docker init` and
are supposed to help with best practices for a python server like we
have here.