Store the openai-key using Google Secret Manager
Before this commit, the OpenAI key was stored in a file that was copied
from the local file system into the container. This commit removes the
need for this local file. Instead, the key is stored in the cloud using
Google Secret Manager, and the value is fetched using the corresponding
API. It works like this:

- A variable `ENV` (dev or prod) is introduced (as an env variable, but
  it is also used in the python application).

- If ENV=dev, authentication to the Secret Manager API is done using
  ADC. The local ADC json file is mounted to the container as a secret.
  See
  https://cloud.google.com/docs/authentication/application-default-credentials#personal.

- If ENV=prod, the name of the secret (openai-key) is passed as a flag
  to the Cloud Run deployment command. A file containing the secret is
  then automatically mounted to the container at a specified path,
  /run/secrets/openai-key.

The `get_openai_api_key` function is updated accordingly. It uses the
`ENV` variable to determine if the value is to be read from
/run/secrets/openai-key or requested from the Secret Manager API.