Rename docker compose server service
Since our repository is called `drivel-server` and the compose server
was called `server` the image name became `drivel-server-server` which
annoyed me a bit. I suggest to rename the service to something like
`backend` instead so the image name becomes `drivel-server-backend`
instead and the container names would be `backend-1` and so on similar
to before. It is just a small OCD fix. When changing name you would want
to run:

```bash
docker compose up --build --remove-orphans
```

the first time you build with the new name.