Add function to get openai secrets
It uses the same logic as fetching the openai key secret. In the prod
environment it reads from a mounted file and for local development it
reads from the google cloud secret manager which requires `gcloud` access.
We right now have two secrets to collect from openai:

- openai api key
- openai organization id

I also added a `asyncio.gather` which starts the fetching of both openai
secrets at the same time so they can run concurrently.