Add testing github workflow
We also needed to add authentication with google cloud to get the needed
secrets. I did this by following this guide:

https://github.com/google-github-actions/auth?tab=readme-ov-file#preferred-direct-workload-identity-federation

The steps for future documentation was:
-  Create a workload identity pool

    ```bash
    gcloud iam workload-identity-pools create "github"
    --project="${PROJECT_ID}" --location="global" --display-name="GitHub
    Actions Pool"
    ```

- Create a Workload Identity Provider in that pool

    ```bash
    gcloud iam workload-identity-pools providers create-oidc "drivel"
    --project="${PROJECT_ID}" --location="global"
    --workload-identity-pool="github" --display-name="My GitHub repo
    Provider"
    --attribute-mapping="google.subject=assertion.sub,attribute.actor=assertion.actor,attribute.repository=assertion.repository,attribute.repository_owner=assertion.repository_owner"
    --attribute-condition="assertion.repository_owner == 'joellidin'"
    --issuer-uri="https://token.actions.githubusercontent.com"
    ```

  Here I am unsure about the attribute mapping and condition but I took
  the one they suggested. There could be some security concerns here so
  we should be careful when adding more people to the project.

- Add the ability to access our secrets

    ```bash gcloud secrets add-iam-policy-binding <secret-name>
    --project="${PROJECT_ID}"
    --role="roles/secretmanager.secretAccessor"
    --member="principalSet://iam.googleapis.com/${WORKLOAD_IDENTITY_POOL_ID}/attribute.repository/${REPO}"\
    ```

    Where
    `WORKLOAD_IDENTITY_POOL_ID=projects/1024792190274/locations/global/workloadIdentityPools/github`
    (created in the step earlier).

Right now, we only have access to the secrets (I think) and if we add
more secrets to the project we need to do this last step in order for
the workflow to work. I am unsure about what we would need to use
`gcloud` for deploying and other kind of actions, but there are more
documentation to read when we are at that stage.

Moreover, since we need the environment variables, I decided to read in
the `.env.template` since all of them are open. And for the python setup
to work I added the `PYTHON_VERSION` variable which we right now have as
3.12.