Add pinned versions of dependencies
It is good since this is an application to keep track of the exact
versions we are using. I have used [`rye`](https://rye-up.com/) for
locking the dependencies as I think it as great tool backed from
astral.sh, the creators of `ruff`. This means however that I added
`tool.rye.dev-dependecies` to the `pyproject.toml`. However, this does
not enforce the usage of `rye`, it just enables it. We can install the
dependencies with pip and use `pip install -r requirements.lock` as I
did in the `Dockerfile`.

To sync the dependencies with `rye` I did:

```sh
rye sync --features "default" --update-all
```