Add `Depends` to STTParameters
Without it, it does not work. I don't understand fully what the
difference is, but it has something to do with dependency injection. But
according to ChatGPT, when using `Depends` you delegate the parsing and
validation of the input to FastAPI, keeping the endpoint function
cleaner and more focused on its primary responsibility.